requirement.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
black==20.8b1
click==7.1.2
decorator==4.4.2
Django==3.1.1
django-cors-headers==3.5.0
django-extensions==3.0.9
django-grappelli==2.14.2
django-ipware==3.0.1
django-jet==1.0.8
django-model-utils==4.0.0
django-redis==4.12.1
djangorestframework==3.11.1
fake-factory==9999.9.9
Faker==4.1.3
graphviz==0.14.1
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2==2.7.7
ptyprocess==0.6.0
pydotplus==2.0.2
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
regex==2020.7.14
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1
